Use tiny bert fixtures in more places
